Reverted form search to be the supervisor portal as well as changed naming.
