+++ app/controllers/main_routes/__init__.py
@@ -23,5 +23,4 @@ from app.controllers.main_routes import download
 from app.controllers.main_routes import laborReleaseForm
 from app.controllers.main_routes import contributors
 from app.controllers.main_routes import studentLaborEvaluation
-from app.controllers.main_routes import formSearch
 from app.controllers.main_routes import search
