+++ app/controllers/main_routes/main_routes.py
@@ -1,34 +1,24 @@
-from flask import flash, send_file
-from app.controllers.main_routes import *
-from app.controllers.main_routes.download import CSVMaker
-from app.login_manager import *
-from app.models.laborStatusForm import LaborStatusForm
-from app.models.department import Department
-from app.models.term import Term
-from app.models.historyType import HistoryType
-from app.models.formHistory import FormHistory
-from datetime import datetime, date
-from flask import request, redirect
-from flask import json, jsonify
-from flask import make_response
-from app.logic.tracy import Tracy
-from app.logic.tracy import InvalidQueryException
-import app.login_manager as login_manager
-import base64
-import time
 import sys
+import operator
+from flask import render_template, request, json, jsonify, redirect, url_for, send_file
+from functools import reduce
+from peewee import JOIN, prefetch
+from app.models.term import Term
+from app.models.department import Department
+from app.models.supervisor import Supervisor
+from app.models.student import Student
+from app.models.laborStatusForm import LaborStatusForm
+from app.models.formHistory import FormHistory
+from app.models.historyType import HistoryType
+from app.models.status import Status
+from app.models.user import User
+from app.models.studentLaborEvaluation import StudentLaborEvaluation
+from app.controllers.admin_routes.allPendingForms import checkAdjustment
+from app.controllers.main_routes import main_bp
+from app.controllers.main_routes.download import CSVMaker
+from app.logic.search import getDepartmentsForSupervisor
+from app.login_manager import require_login, logout
 
-currentlyEnrolledBNumbers = []
-
-# Check if a student is currently a student at Berea. Only get the list from Tracy once
-# This means we'll need to restart the application to refresh tracy data (currently a nightly restart)
-def isCurrentStudent(bnumber):
-    global currentlyEnrolledBNumbers
-
-    if not currentlyEnrolledBNumbers:
-        currentlyEnrolledBNumbers = [s.ID.strip() for s in Tracy().getStudents()]
-
-    return (bnumber in currentlyEnrolledBNumbers)
 
 @main_bp.before_app_request
 def before_request():
@@ -36,249 +26,324 @@ def before_request():
 
 @main_bp.route('/logout', methods=['GET'])
 def logout():
-    return redirect(login_manager.logout())
+    return redirect(logout())
+
+# Global variable that will store the query result.
+# It is made global to be used later in creating CSV file.
+formSearchResults = None
+sleJoin = False
 
 @main_bp.route('/', methods=['GET', 'POST'])
-@main_bp.route('/main/students', methods=['GET', 'POST'])
-@main_bp.route('/main/department', methods=['GET', 'POST'])
-@main_bp.route('/main/department/<department>', methods=['GET', 'POST'])
-def index(department = None):
-    try:
-        currentUser = require_login()
-        if not currentUser:
-            return render_template('errors/403.html'), 403
-        if not currentUser.isLaborAdmin:
-            if currentUser.student and not currentUser.supervisor:   # logged in as a student
-                return redirect('/laborHistory/' + currentUser.student.ID)
-            if currentUser.supervisor:       # logged in as a Supervisor
-                # Checks all the forms where the current user has been the creator or the supervisor, and grabs all the departments associated with those forms. Will only grab each department once.
-                departments = FormHistory.select(FormHistory.formID.department.DEPT_NAME) \
-                                .join_from(FormHistory, LaborStatusForm) \
-                                .join_from(LaborStatusForm, Department) \
-                                .where((FormHistory.formID.supervisor == currentUser.supervisor.ID) | (FormHistory.createdBy == currentUser)) \
-                                .order_by(FormHistory.formID.department.DEPT_NAME.asc()) \
-                                .distinct()
+def supervisorPortal():
+    '''
+    When the request is GET the function populates the General Search interface dropdown menus with their corresponding values.
+    If the request is POST it also populates the datatable with data based on user input.
+    '''
+    currentUser = require_login()
+    if not currentUser or not currentUser.supervisor:
+        return render_template('errors/403.html'), 403
 
-        else:   # logged in as an admin
-            # Grabs every single department that currently has at least one labor status form in it
-            departments = FormHistory.select(FormHistory.formID.department.DEPT_NAME) \
-                            .join_from(FormHistory, LaborStatusForm) \
-                            .join_from(LaborStatusForm, Department) \
-                            .order_by(FormHistory.formID.department.DEPT_NAME.asc()) \
-                            .distinct()
+    terms = LaborStatusForm.select(LaborStatusForm.termCode).distinct().order_by(LaborStatusForm.termCode.desc())
 
-        todayDate = date.today()
-        # Grabs all the labor status forms where the current user is the supervisor
-        formsBySupervisees = []
-        if currentUser.supervisor:
-            formsBySupervisees = FormHistory.select().join_from(FormHistory, LaborStatusForm).join_from(FormHistory, HistoryType).where(FormHistory.formID.supervisor == currentUser.supervisor.ID,
-            FormHistory.historyType.historyTypeName == "Labor Status Form").order_by(FormHistory.formID.startDate.desc())
-            formsBySupervisees = sorted(formsBySupervisees,key=lambda f:f.reviewedDate if f.reviewedDate else f.createdDate, reverse=True)
+    if currentUser.isLaborAdmin or currentUser.isFinancialAidAdmin or currentUser.isSaasAdmin:
+        departments = Department.select().order_by(Department.DEPT_NAME.asc())
+        departments = [department for department in departments]
+        supervisors = Supervisor.select().order_by(Supervisor.FIRST_NAME.asc())
+        students = Student.select().order_by(Student.FIRST_NAME.asc())
 
-        inactiveSupervisees = []
-        currentSupervisees = []
-        pastSupervisees = []
+    else:
 
-        tic = time.perf_counter()
-        for supervisee in formsBySupervisees: # go through all the form in the formsBySupervisees
-            student_processed = False # whether or not the student has been added to the list
+        departments = list(getDepartmentsForSupervisor(currentUser))
 
-            if isCurrentStudent(supervisee.formID.studentSupervisee.ID):
-                for student in currentSupervisees:
-                    # Checks whether student has already been added as an current student.
-                    if (supervisee.formID.studentSupervisee.ID) == (student.formID.studentSupervisee.ID):
-                        student_processed = True
-                for student in pastSupervisees:
-                    # Checks whether student has already been added as an past student.
-                    if (supervisee.formID.studentSupervisee.ID) == (student.formID.studentSupervisee.ID):
-                        student_processed = True
+        # convert department objects to strings
+        departments = [department for department in departments]
 
-                # If a student has not yet been added to the view, they are appended as an active student.
-                if student_processed == False:
-                    if supervisee.formID.endDate < todayDate:
-                        pastSupervisees.append(supervisee)
-                    elif supervisee.formID.endDate >= todayDate:
-                        studentFormHistory = FormHistory.select().where(FormHistory.formID == supervisee.formID.laborStatusFormID).order_by(FormHistory.formHistoryID.desc())[0]
-                        if studentFormHistory.historyType.historyTypeName == "Labor Release Form":
-                            if studentFormHistory.status.statusName == "Approved":
-                                pastSupervisees.append(supervisee)
-                            else:
-                                currentSupervisees.append(supervisee)
-                        else:
-                            currentSupervisees.append(supervisee)
+        supervisors = (Supervisor.select()
+                            .join_from(Supervisor, LaborStatusForm)
+                            .join_from(LaborStatusForm, Department)
+                            .where(Department.DEPT_NAME.in_(departments))
+                            .distinct())
 
-            else: # if they are inactive
-                for student in inactiveSupervisees:
-                    # Checks whether student has already been added as an active student.
-                    if (supervisee.formID.studentSupervisee.ID) == (student.formID.studentSupervisee.ID):
-                        student_processed = True
-                # If a student has not yet been added to the view, they are appended as an active student.
-                if student_processed == False:
-                    inactiveSupervisees.append(supervisee)
+        students = (Student.select()
+                    .join_from(Student, LaborStatusForm)
+                    .join_from(LaborStatusForm, Department)
+                    .where(Department.DEPT_NAME.in_(departments))
+                    .distinct())
+    if request.method == 'POST':
+        return getDatatableData(request)
 
-        toc = time.perf_counter()
-        # print("Processed {} supervisor students in {:0.4f} seconds".format(len(formsBySupervisees), toc-tic))
+    return render_template('main/supervisorPortal.html',
+                            terms = terms,
+                            supervisors = supervisors,
+                            students = students,
+                            departments = departments,
+                            department = None,
+                            currentUser = currentUser
+                            )
 
-        # On the click of the download button, 'POST' method will send all checked boxes from modal to excel maker
-        if request.method == 'POST':
-            value =[]
-            # The "Try" and "Except" block here is needed because if the user tries to use the download button before they chose
-            # a department from the Department dropdown, it will throw a NameError. The reason behind the error is because the vairbales
-            # "currentDepartmentStudents", "allDepartmentStudents", and "inactiveDepStudent" are empty until the user chooses a department, so
-            # trying to iterate through the empty variables causes the error. The "Try" and "Except" blocks will catch this error so that
-            # a user can use the download button before they chose a department.
-            try:
-                for form in currentDepartmentStudents:
-                    name = str(form.formID.laborStatusFormID)
-                    if request.form.get(name):
-                        value.append(request.form.get(name))
-                for form in allDepartmentStudents:
-                    name = str(form.formID.laborStatusFormID)
-                    if request.form.get(name):
-                        value.append(request.form.get(name))
-                for form in inactiveDepStudent:
-                    name = str(form.formID.laborStatusFormID)
-                    if request.form.get(name):
-                        value.append(request.form.get(name))
-            except NameError as e:
-                print("The runtime error happens because a department has not yet been selected.")
-            for form in currentSupervisees:
-                name = str(form.formID.laborStatusFormID)
-                if request.form.get(name):
-                    value.append( request.form.get(name))
-            for form in pastSupervisees:
-                name = str(form.formID.laborStatusFormID)
-                if request.form.get(name):
-                    value.append( request.form.get(name))
-            for form in inactiveSupervisees:
-                name = str(form.formID.laborStatusFormID)
-                if request.form.get(name):
-                    value.append( request.form.get(name))
+def getDatatableData(request):
+    '''
+    This function runs a query based on selected options in the front-end and retrieves the appropriate forms.
+    Then, it puts all the retrieved data in appropriate form to be send to the ajax call in the supervisorPortal.js file.
+    '''
 
+    # 'draw', 'start', 'length', 'order[0][column]', 'order[0][dir]' are built-in parameters, i.e.,
+    # they are implicitly passed as part of the AJAX request when using datatable server-side processing
+    draw = int(request.form.get('draw', -1))
+    rowNumber = int(request.form.get('start', -1))
+    rowsPerPage = int(request.form.get('length', -1))
+    sortColIndex = int(request.form.get("order[0][column]", -1))
+    order = request.form.get('order[0][dir]')
+    queryFilterData = request.form.get('data')
+    queryFilterDict = json.loads(queryFilterData)
+    # Dictionary to match column indices with column names in the DB
+    # It is used for identifying the column that needs to be sorted
+    colIndexColNameMap = {  0: Term.termCode,
+                            1: Department.DEPT_NAME,
+                            2: Supervisor.FIRST_NAME,
+                            3: Student.FIRST_NAME,
+                            4: LaborStatusForm.POSN_CODE,
+                            5: LaborStatusForm.weeklyHours,
+                            6: LaborStatusForm.startDate,
+                            7: User.username,
+                            8: FormHistory.status,
+                            9: FormHistory.historyType,
+                            10: StudentLaborEvaluation.ID}
 
-            # Prevents 'POST' method from sending the same value twice to excel maker
+    termCode = queryFilterDict.get('termCode', "")
+    departmentId = queryFilterDict.get('departmentID', "")
+    supervisorId = queryFilterDict.get('supervisorID', "")
+    studentId = queryFilterDict.get('studentID', "")
+    formStatusList = queryFilterDict.get('formStatus', "") # form status radios
+    formTypeList = queryFilterDict.get('formType', "") # form type radios
+    evaluationStatus = queryFilterDict.get('evaluations', "") # evaluation radios
 
-            excel = CSVMaker("studentList", value, includeEvals = True)
-            # Returns the file path so the button will download the file
-            return send_file(excel.relativePath,as_attachment=True, attachment_filename=excel.relativePath.split('/').pop())
+    fieldValueMap = {Term.termCode: termCode,
+                     Department.departmentID: departmentId,
+                     Student.ID: studentId,
+                     Supervisor.ID: supervisorId,
+                     FormHistory.status: formStatusList,
+                     FormHistory.historyType: formTypeList,
+                     StudentLaborEvaluation.ID: evaluationStatus}
 
-        return render_template( 'main/index.html',
-    				    title=('Home'),
-                        currentSupervisees = currentSupervisees,
-                        pastSupervisees = pastSupervisees,
-                        inactiveSupervisees = inactiveSupervisees,
-                        UserID = currentUser,
-                        currentUserDepartments = departments,
-                        department = department
-                              )
-    except Exception as e:
-        #TODO We have to return some sort of error page
-        print('Error in Supervisor Portal:', e)
-        return "",500
+    clauses = []
 
-@main_bp.route('/main/department/selection/<departmentSelected>', methods=['GET'])
-def populateDepartment(departmentSelected):
-    try:
-        currentUser = require_login()
-        todayDate = date.today()
-
-        try:
-            department = Department.get(Department.DEPT_NAME == departmentSelected)
-        except DoesNotExist:
-            print("Department '{}' does not exist".format(departmentSelected))
-            return jsonify({"Success": False})
-
-        deptPositions = [p.POSN_CODE for p in Tracy().getPositionsFromDepartment(department.ORG, department.ACCOUNT)]
-
-        # This will retrieve all the forms that are tied to the department the user selected from the select picker
-        formsByDept = FormHistory.select().join_from(FormHistory, LaborStatusForm).join_from(FormHistory, HistoryType).where((FormHistory.historyType.historyTypeName == "Labor Status Form") & (FormHistory.formID.POSN_CODE << deptPositions)).order_by(FormHistory.formID.endDate.desc())
-
-        # These variables need to be global variables because they need to be used in other requests
-        global currentDepartmentStudents
-        global allDepartmentStudents
-        global inactiveDepStudent
-
-        currentDepartmentStudents = []
-        allDepartmentStudents = []
-        inactiveDepStudent = []
-
-        tic = time.perf_counter()
-        for supervisee in formsByDept: # go through all the form in the formsBySupervisees
-            student_processed = False  # This variable dictates whether a student has already been added to the supervisor's portal
-
-            if isCurrentStudent(supervisee.formID.studentSupervisee.ID):
-                for student in currentDepartmentStudents:
-                    if (supervisee.formID.studentSupervisee.ID) == (student.formID.studentSupervisee.ID):  # Checks whether student has already been added as an current student.
-                        student_processed = True
-                for student in allDepartmentStudents:
-                    if (supervisee.formID.studentSupervisee.ID) == (student.formID.studentSupervisee.ID):  # Checks whether student has already been added as an past student.
-                        student_processed = True
-                if student_processed == False:  # If a student has not yet been added to the view, they are appended as an active student.
-                    # If a forms "endDate" is less than today's date, then we know the form is from the past and will go into "All Departments"
-                    if supervisee.formID.endDate < todayDate:
-                        allDepartmentStudents.append(supervisee)
-                    # If a forms "endDate" is greater than today's date, then we need to look at other conditions
-                    elif supervisee.formID.endDate >= todayDate:
-                        # For every form at this point, we need to see if there are any "Labor Release Forms" tied to the form
-                        studentFormHistory = FormHistory.select().where(FormHistory.formID == supervisee.formID.laborStatusFormID).order_by(FormHistory.formHistoryID.desc())[0]
-                        if studentFormHistory.historyType.historyTypeName == "Labor Release Form":
-                            # If the form has a "Labor Release Form" with a status of "Approved", then we know the student is no longer employed
-                            if studentFormHistory.status.statusName == "Approved":
-                                allDepartmentStudents.append(supervisee)
-                            # If the form has a "Labor Release Form" with a status not equal to "Approved", then we know the student is still employed
-                            else:
-                                currentDepartmentStudents.append(supervisee)
-                        # If the form does not have a "Labor Release Form", then we know the student is still employed
-                        else:
-                            currentDepartmentStudents.append(supervisee)
+    global sleJoin
+    # WHERE clause conditions are dynamically generated using model fields and selectpicker values
+    for field, value in fieldValueMap.items():
+        if value != "" and value:
+            # "is" is used to compare the two peewee objects as opposed to "==" operator.
+            if field is FormHistory.historyType:
+                for val in value:
+                    clauses.append(field == val)
+            elif field is FormHistory.status:
+                for val in value:
+                    clauses.append(field == val)
+            elif field is StudentLaborEvaluation.ID:
+                sleJoin = value[0]       # LSF exists but SLE does not (LOJ)
             else:
-                for student in inactiveDepStudent:
-                    if (supervisee.formID.studentSupervisee.ID) == (student.formID.studentSupervisee.ID):  # Checks whether student has already been added as an active student.
-                        student_processed = True
-                if student_processed == False:  # If a student has not yet been added to the view, they are appended as an active student.
-                    inactiveDepStudent.append(supervisee)
+                clauses.append(field == value)
 
-        toc = time.perf_counter()
-        print("Processed {} department forms in {:0.4f} seconds".format(len(formsByDept),toc - tic))
+    # This expression creates SQL AND operator between the conditions added to 'clauses' list
+    expression = reduce(operator.and_, clauses)
+
+    global formSearchResults
+    formSearchResults = (FormHistory.select()
+                        .join(LaborStatusForm, on=(FormHistory.formID == LaborStatusForm.laborStatusFormID))
+                        .join(Department, on=(LaborStatusForm.department == Department.departmentID))
+                        .join(Supervisor, on=(LaborStatusForm.supervisor == Supervisor.ID))
+                        .join(Student, on=(LaborStatusForm.studentSupervisee == Student.ID))
+                        .join(Term, on=(LaborStatusForm.termCode == Term.termCode))
+                        .join(User, on=(FormHistory.createdBy == User.userID))
+                        .where(expression))
+
+    if sleJoin:
+        if sleJoin == "evalMidyearMissing" or sleJoin == "evalMidyearComplete":
+            #grab all the midyear evaluationStatus
+            evalResults = StudentLaborEvaluation.select(StudentLaborEvaluation.formHistoryID).where(StudentLaborEvaluation.formHistoryID.formID.termCode == termCode, StudentLaborEvaluation.is_midyear_evaluation == True, StudentLaborEvaluation.is_submitted == True)
+        else:
+            #grab all the final evaluationStatus
+            evalResults = StudentLaborEvaluation.select(StudentLaborEvaluation.formHistoryID).where(StudentLaborEvaluation.formHistoryID.formID.termCode == termCode, StudentLaborEvaluation.is_midyear_evaluation == False, StudentLaborEvaluation.is_submitted == True)
+        if sleJoin == "evalMidyearMissing":
+            formSearchResults = formSearchResults.select().where(FormHistory.formHistoryID.not_in(evalResults))
+        elif sleJoin == "evalMidyearComplete":
+            formSearchResults = formSearchResults.select().where(FormHistory.formHistoryID.in_(evalResults))
+        elif sleJoin == "evalMissing":
+            formSearchResults = formSearchResults.select().where(FormHistory.formHistoryID.not_in(evalResults))
+        elif sleJoin == "evalComplete":
+            formSearchResults = formSearchResults.select().where(FormHistory.formHistoryID.in_(evalResults))
+
+    recordsTotal = formSearchResults.count()
+
+    # Sorting a column in descending order when a specific column is chosen
+    # Initially, it sorts by the Term column as specified in supervisorPortal.js
+    if order == "desc":
+        filteredSearchResults = formSearchResults.order_by(-colIndexColNameMap[sortColIndex]).limit(rowsPerPage).offset(rowNumber)
+    # Sorting a column in ascending order when a specific column is chosen
+    else:
+        filteredSearchResults = formSearchResults.order_by(colIndexColNameMap[sortColIndex]).limit(rowsPerPage).offset(rowNumber)
+
+    formattedData = getFormattedData(filteredSearchResults)
+    formsDict = {"draw": draw, "recordsTotal": recordsTotal, "recordsFiltered": recordsTotal, "data": formattedData}
+
+    return jsonify(formsDict)
+
+def getFormattedData(filteredSearchResults):
+    '''
+    Putting the data in the correct format to be used by the JS file.
+    Because this implementation is using server-side processing of datatables,
+    the HTML for the datatables are also formatted here.
+    '''
+
+    supervisorStudentHTML = '<a href="#" class="hover_indicator" aria-label="{}">{} </a><a href="mailto:{}"><span class="glyphicon glyphicon-envelope mailtoIcon"></span></a>'
+    departmentHTML = '<a href="#" class="hover_indicator" aria-label="{}-{}"> {}</a>'
+    positionHTML = '<a href="#" class="hover_indicator" aria-label="{}"> {}</a>'
+    formattedData = []
+    for form in filteredSearchResults:
+        # The order in which you append the items to 'record' matters and it should match the order of columns on the table!
+        record = []
+        # Term
+        record.append(form.formID.termCode.termName)
+        # Department
+        record.append(departmentHTML.format(
+              form.formID.department.ORG,
+              form.formID.department.ACCOUNT,
+              form.formID.department.DEPT_NAME))
+        # Supervisor
+        supervisorField = supervisorStudentHTML.format(
+                            form.formID.supervisor.ID,
+                            f'{form.formID.supervisor.FIRST_NAME} {form.formID.supervisor.LAST_NAME}',
+                            form.formID.supervisor.EMAIL)
+        # Position
+        positionField = positionHTML.format(
+                        form.formID.POSN_TITLE,
+                        f'{form.formID.POSN_CODE} ({form.formID.WLS})')
+        # Hours
+        hoursField = form.formID.weeklyHours if form.formID.weeklyHours else form.formID.contractHours
+
+        # Adjustment Form Specific Data
+        checkAdjustment(form)
+        if (form.adjustedForm):
+            if form.adjustedForm.fieldAdjusted == "supervisor":
+                newSupervisor = supervisorStudentHTML.format(
+                                form.adjustedForm.oldValue['ID'],
+                                form.adjustedForm.newValue,
+                                form.adjustedForm.oldValue['email'])
+                supervisorField = f'<s aria-label="true">{supervisorField}</s><br>{newSupervisor}'
+
+            if form.adjustedForm.fieldAdjusted == "position":
+                newPosition = positionHTML.format(
+                              form.adjustedForm.oldValue,
+                              form.adjustedForm.newValue)
+                positionField = f'<s aria-label="true">{positionField}</s><br>{newPosition}'
+
+            if form.adjustedForm.fieldAdjusted == "weeklyHours"  or  form.adjustedForm.fieldAdjusted == "contractHours":
+                newHours = form.adjustedForm.newValue
+                hoursField = f'<s aria-label="true">{hoursField}</s><br>{newHours}'
+
+        record.append(supervisorField)
+        # Student
+        record.append(supervisorStudentHTML.format(
+              form.formID.studentSupervisee.ID,
+              f'{form.formID.studentSupervisee.FIRST_NAME} {form.formID.studentSupervisee.LAST_NAME}',
+              form.formID.studentSupervisee.STU_EMAIL))
+
+        record.append(f'{positionField}<br>{form.formID.jobType}')
+        record.append(hoursField)
+        # Contract Dates
+        record.append("<br>".join([form.formID.startDate.strftime('%m/%d/%y'),
+                                   form.formID.endDate.strftime('%m/%d/%y')]))
+        # Created By
+        record.append(supervisorStudentHTML.format(
+              form.createdBy.supervisor.ID if form.createdBy.supervisor else form.createdBy.student.ID,
+              form.createdBy.username,
+              form.createdBy.email,
+              form.createdDate.strftime('%m/%d/%y')))
+
+        # Form Status
+        record.append(form.status.statusName)
+        # Form Type
+        formTypeNameMapping = {
+            "Labor Status Form": "Original",
+            "Labor Adjustment Form": "Adjusted",
+            "Labor Overload Form": "Overload",
+            "Labor Release Form": "Release"}
+        originalFormTypeName = form.historyType.historyTypeName
+        mappedFormTypeName = formTypeNameMapping[originalFormTypeName]
+        record.append(mappedFormTypeName)
+
+        # Evaluation status
+        # TODO Skipping adding to the table. Requires database work to get SLE out from form (formHistory, to be precise)
+
+        laborHistoryId = form.formHistoryID
+        laborStatusFormId = form.formID.laborStatusFormID
+        actionsButton = getActionButtonLogic(form, laborHistoryId, laborStatusFormId)
+
+        record.append(actionsButton)
+        formattedData.append(record)
+
+    return formattedData
 
 
-        # This section will format our JSON data with the key-value pairs we want to pass back to the AJAX call in our JS file.
-        departmentStudents = []
-        for i in currentDepartmentStudents:
-            departmentStudents.append({"Status":"Current Department Students",
-                                        "Student": i.formID.studentSupervisee.FIRST_NAME + " " + i.formID.studentSupervisee.LAST_NAME,
-                                        "BNumber": i.formID.studentSupervisee.ID,
-                                        "Term": i.formID.termCode.termName,
-                                        "Position": i.formID.POSN_TITLE,
-                                        "checkboxModalClass" : "currentDepartmentModal",
-                                        "activeStatus" : "True",
-                                        "formID" : i.formID.laborStatusFormID,
-                                        "Department": i.formID.department.DEPT_NAME,
-                                        "formStatus": i.status.statusName})
-        for i in allDepartmentStudents:
-            departmentStudents.append({"Status":"All Department Students",
-                                        "Student": i.formID.studentSupervisee.FIRST_NAME + " " + i.formID.studentSupervisee.LAST_NAME,
-                                        "BNumber": i.formID.studentSupervisee.ID,
-                                        "Term": i.formID.termCode.termName,
-                                        "Position": i.formID.POSN_TITLE,
-                                        "checkboxModalClass" : "allDepartmentModal",
-                                        "activeStatus" : "True",
-                                        "formID" : i.formID.laborStatusFormID,
-                                        "Department": i.formID.department.DEPT_NAME,
-                                        "formStatus": "Past Student"})
-        for i in inactiveDepStudent:
-            departmentStudents.append({"Status":"All Department Students",
-                                    "Student": i.formID.studentSupervisee.FIRST_NAME + " " + i.formID.studentSupervisee.LAST_NAME,
-                                    "BNumber": i.formID.studentSupervisee.ID,
-                                    "Term": i.formID.termCode.termName,
-                                    "Position": i.formID.POSN_TITLE,
-                                    "checkboxModalClass" : "allDepartmentModal",
-                                    "activeStatus" : "False",
-                                    "formID" : i.formID.laborStatusFormID,
-                                    "Department": i.formID.department.DEPT_NAME,
-                                    "formStatus": i.status.statusName})
-        return json.dumps(departmentStudents)
+def getActionButtonLogic(form, laborHistoryId, laborStatusFormId):
+    '''
+    This function determines the options shown on the Actions dropdown, which depends on form type and form status.
+    '''
 
-    except Exception as e:
-        print('ERROR in Department Students:', e)
-        return jsonify({"Success": False})
+    actionsButtonDropdownHTML = '<div class="dropdown"><button class="btn btn-primary dropdown-toggle" type="button" id="menu1" data-toggle="dropdown">Actions <span class="caret"></span></button>' +\
+                                '<ul class="dropdown-menu" role="menu" aria-labelledby="menu1" style="min-width: 100%;">{}{}{}{}{}</ul></div>'
+    actionsListHTML = '<li role="presentation"><a role="menuitem" href="{}">{}</a></li>'
+    manageOptionHTML = actionsListHTML.format('#', '<span id="{}" onclick="{}">{}</span>')
+    denyApproveNotesOptionsHTML = actionsListHTML.format('#', '<span id="{}" onclick="{}" data-toggle="modal" data-target="#{}">{}</span>')
+    modifyOptionHTML = actionsListHTML.format('/alterLSF/{lsfID}', '<span id="edit_{lsfID}">Modify</span>')
+
+    # Actions button and its options
+    actionsButton = ""
+    approve = ""
+    deny = ""
+    manage = ""
+    modify = ""
+    notes = ""
+
+    if form.historyType.historyTypeName == "Labor Status Form":
+        manage = manageOptionHTML.format(laborHistoryId, f"loadLaborHistoryModal({laborStatusFormId})", 'Labor History')
+        actionsButton = actionsButtonDropdownHTML.format(manage, approve, deny, modify, notes)
+
+    if form.status.statusName in ("Pending", "Pre-Student Approval"):
+        # show notes modal for all pending labor status forms
+        notes = denyApproveNotesOptionsHTML.format(f'notes_{laborHistoryId}', f'getNotes({laborStatusFormId})', 'NotesModal', 'View Notes')
+        if form.overloadForm:
+            manage = manageOptionHTML.format(laborHistoryId, f"loadOverloadModal({laborHistoryId}, {laborStatusFormId})", 'Manage')
+        elif form.releaseForm:
+            manage = manageOptionHTML.format(laborHistoryId, f"loadReleaseModal({laborHistoryId}, {laborStatusFormId})", 'Manage')
+        elif form.adjustedForm:
+            deny = denyApproveNotesOptionsHTML.format(f"reject_{laborHistoryId}", f"insertDenial({laborHistoryId})", 'denyModal', 'Deny')
+            approve = denyApproveNotesOptionsHTML.format("", f"insertApprovals({laborHistoryId})", 'approvalModal', 'Approve')
+        else: # if it is the original labor status form
+            modify = modifyOptionHTML.format(lsfID = laborStatusFormId)
+            deny = denyApproveNotesOptionsHTML.format(f"reject_{laborHistoryId}", f"insertDenial({laborHistoryId})", 'denyModal', 'Deny')
+            approve = denyApproveNotesOptionsHTML.format("", f"insertApprovals({laborHistoryId});", 'approvalModal', 'Approve')
+        actionsButton = actionsButtonDropdownHTML.format(manage, approve, deny, modify, notes)
+
+    return actionsButton
+
+
+@main_bp.route('/supervisorPortal/download', methods=['POST'])
+def downloadSupervisorPortalResults():
+    '''
+    This function uses the general search results, stored in a global variable, to
+    generate a CSV file of datatable data.
+    '''
+
+    global formSearchResults
+    global sleJoin
+    if sleJoin == "evalComplete":
+        includeEvals = "Final"
+    elif sleJoin == "evalMidyearComplete":
+        includeEvals = "Midyear"
+    else:
+        includeEvals = False
+
+    formSearchResults = formSearchResults.order_by(-FormHistory.createdDate)
+    excel = CSVMaker("supervisorPortal", formSearchResults, includeEvals = includeEvals)
+    return send_file(excel.relativePath, as_attachment=True, attachment_filename=excel.relativePath.split('/').pop())
