--- app/static/css/index.css
@@ -1,54 +1,0 @@
-#portalTitle{
-  font-size: 24px;
-}
-#myInput {
-  background-image: url('/css/searchicon.png'); /* Add a search icon to input */
-  background-position: 10px 12px; /* Position the search icon */
-  background-repeat: no-repeat; /* Do not repeat the icon image */
-  /*float: right; /* places search bar on the right */
-  width: 15%; /* small width, button size */
-  font-size: 16px; /* Increase font-size */
-  padding: 6px 10px 6px ; /* Add some padding */
-  border: 1px solid #ddd; /* Add a grey border */
-  margin-bottom: 12px; /* Add some space below the input */
-  right: 0;
-}
-
-.pushReft {
-  padding-left: 10px;
-}
-
-.pushRight {
-  float: right;
-}
-
-.student_table th, .student_table td {
-  text-align: left; /* Left-align text */
-  padding: 12px; /* Add padding */
-}
-
-.student_table tr {
-  /* Add a bottom border to all table rows */
-  border-bottom: 1px solid #ddd;
-}
-
-.student_table tr.header, .student_table tr:hover {
-  /* Add a grey background color to the table header and on hover */
-  background-color: #f1f1f1;
-}
-iframe{
-  display: none;
-}
-
-#closeButton {
-  margin-right: 405px
-}
-
-#download {
-  margin-bottom: 10px;
-  display: none;
-}
-
-#studentList {
-  display: none;
-}
